First steps on new game scraper
Currently, game scraper loops through accts that are already in the
list.

Checking for the time condition seems to be working. Though I'm seeing
something weird: the new games being scanned are AFTER the most recent
game and BEFORE 2016-06-19 (which is when I think Tyler scraped last).

changed calculate elo, wanted to save the dict for the new scraper.